Move more of managing course to own components
