+++ services/main-frontend/src/components/page-specific/manage/courses/id/pages/ChapterImageWidget.tsx
@@ -1,11 +1,10 @@
 import React, { useState } from "react"
 import { useTranslation } from "react-i18next"
 
-import { removeChapterImage, setChapterImage } from "../services/backend/chapters"
-import { Chapter } from "../shared-module/bindings"
-import Button from "../shared-module/components/Button"
-
-import UploadImageForm from "./forms/UploadImageForm"
+import { removeChapterImage, setChapterImage } from "../../../../../../services/backend/chapters"
+import { Chapter } from "../../../../../../shared-module/bindings"
+import Button from "../../../../../../shared-module/components/Button"
+import UploadImageForm from "../../../../../forms/UploadImageForm"
 
 export interface ChapterImageControlsProps {
   chapter: Chapter
