+++ services/main-frontend/src/pages/manage/courses/[id]/new-course-instance.tsx
@@ -1,14 +1,14 @@
 import { css } from "@emotion/css"
 import { useRouter } from "next/router"
-import React, { useState } from "react"
+import React from "react"
 import { useTranslation } from "react-i18next"
 import { useMutation } from "react-query"
 
 import Layout from "../../../../components/Layout"
-import Form from "../../../../components/forms/CourseInstanceForm"
+import NewCourseInstanceForm from "../../../../components/page-specific/manage/courses/id/new-course-instance/NewCourseInstanceForm"
 import { newCourseInstance } from "../../../../services/backend/courses"
 import { CourseInstanceForm } from "../../../../shared-module/bindings"
-import { isErrorResponse } from "../../../../shared-module/bindings.guard"
+import ErrorBanner from "../../../../shared-module/components/ErrorBanner"
 import { withSignedIn } from "../../../../shared-module/contexts/LoginStateContext"
 import { frontendWideWidthCenteredComponentStyles } from "../../../../shared-module/styles/componentStyles"
 import {
@@ -26,24 +26,15 @@ const NewCourseInstance: React.FC<Props> = ({ query }) => {
   const courseId = query.id
 
   const router = useRouter()
-  const [error, setError] = useState<string | null>(null)
 
   const mutation = useMutation(
     async (form: CourseInstanceForm) => {
-      setError(null)
       await newCourseInstance(courseId, form)
     },
     {
       onSuccess: () => {
         // eslint-disable-next-line i18next/no-literal-string
         router.push(`/manage/courses/${courseId}`)
-      },
-      onError: (err) => {
-        if (isErrorResponse(err)) {
-          setError(t("message-creating-failed"))
-        } else {
-          setError(t("message-creating-failed"))
-        }
       },
     },
   )
@@ -57,8 +48,8 @@ const NewCourseInstance: React.FC<Props> = ({ query }) => {
         `}
       >
         <h1>{t("new-course-instance")}</h1>
-        {error && <div>{error}</div>}
-        <Form
+        {mutation.isError && <ErrorBanner variant={"readOnly"} error={mutation.error} />}
+        <NewCourseInstanceForm
           initialData={null}
           onSubmit={(data) => mutation.mutate(data)}
           onCancel={() => router.back()}
