+++ services/main-frontend/src/pages/manage/courses/[id]/pages.tsx
@@ -5,13 +5,16 @@ import React, { useState } from "react"
 import { useTranslation } from "react-i18next"
 import { useQuery } from "react-query"
 
-import ChapterImageWidget from "../../../../components/ChapterImageWidget"
 import Layout from "../../../../components/Layout"
-import NewChapterForm from "../../../../components/forms/NewChapterForm"
-import PageList from "../../../../components/lists/PageList"
+import ChapterImageWidget from "../../../../components/page-specific/manage/courses/id/pages/ChapterImageWidget"
+import ManageCourseStructure from "../../../../components/page-specific/manage/courses/id/pages/ManageCourseStructure"
+import NewChapterForm from "../../../../components/page-specific/manage/courses/id/pages/NewChapterForm"
+import PageList from "../../../../components/page-specific/manage/courses/id/pages/PageList"
 import { fetchCourseStructure } from "../../../../services/backend/courses"
 import Button from "../../../../shared-module/components/Button"
 import DebugModal from "../../../../shared-module/components/DebugModal"
+import ErrorBanner from "../../../../shared-module/components/ErrorBanner"
+import Spinner from "../../../../shared-module/components/Spinner"
 import { withSignedIn } from "../../../../shared-module/contexts/LoginStateContext"
 import { normalWidthCenteredComponentStyles } from "../../../../shared-module/styles/componentStyles"
 import dontRenderUntilQueryParametersReady, {
@@ -24,30 +27,8 @@ export interface CoursePagesProps {
 }
 
 const CoursePages: React.FC<CoursePagesProps> = ({ query }) => {
-  const { t } = useTranslation()
   const { id } = query
-  const { isLoading, error, data, refetch } = useQuery(`course-structure-${id}`, () =>
-    fetchCourseStructure(id),
-  )
-  const [showForm, setShowForm] = useState(false)
-
-  if (error) {
-    return <div>{t("error-title")}</div>
-  }
-
-  if (isLoading || !data) {
-    return <div>{t("loading-text")}</div>
-  }
-
-  const handleCreateChapter = async () => {
-    setShowForm(!showForm)
-    await refetch()
-  }
-
-  // eslint-disable-next-line i18next/no-literal-string
-  const pagesByChapter = groupBy(data.pages, "chapter_id")
-
-  const maxPart = max(data.chapters.map((p) => p.chapter_number))
+  const getCourseStructure = useQuery(`course-structure-${id}`, () => fetchCourseStructure(id))
 
   return (
     <Layout navVariant="complex">
@@ -57,67 +38,15 @@ const CoursePages: React.FC<CoursePagesProps> = ({ query }) => {
           margin-bottom: 1rem;
         `}
       >
-        <h1>{t("course-overview-for", { "course-name": data.course.name })}</h1>
-        <h2>{t("pages")}</h2>
-        <PageList
-          data={data.pages.filter((page) => !page.chapter_id)}
-          refetch={refetch}
-          courseId={data.course.id}
-        />
-        <h2>{t("chapters")}</h2>
-        <div>
-          {data.chapters
-            .filter((chapter) => !chapter.deleted_at)
-            .sort((a, b) => a.chapter_number - b.chapter_number)
-            .map((chapter) => (
-              <div
-                className={css`
-                  border: 1px solid black;
-                  padding: 2rem;
-                  margin-bottom: 1rem;
-                `}
-                key={chapter.id}
-              >
-                <h3>
-                  {t("title-chapter", {
-                    "chapter-number": chapter.chapter_number,
-                    "chapter-name": chapter.name,
-                  })}
-                </h3>
-                <ChapterImageWidget chapter={chapter} onChapterUpdated={() => refetch()} />
-                <PageList
-                  data={pagesByChapter[chapter.id] ?? []}
-                  refetch={refetch}
-                  courseId={data.course.id}
-                  chapter={chapter}
-                />
-              </div>
-            ))}
-
-          <Button variant="primary" size="medium" onClick={() => setShowForm(!showForm)}>
-            {t("button-text-new")}
-          </Button>
-
-          <Dialog open={showForm} onClose={() => setShowForm(!showForm)}>
-            <div
-              className={css`
-                margin: 1rem;
-              `}
-            >
-              <Button variant="primary" size="medium" onClick={() => setShowForm(!showForm)}>
-                {t("button-text-close")}
-              </Button>
-              <NewChapterForm
-                courseId={data.course.id}
-                onSubmitForm={handleCreateChapter}
-                chapterNumber={(maxPart ?? 0) + 1}
-              />
-            </div>
-          </Dialog>
-        </div>
+        {getCourseStructure.isError && <ErrorBanner variant={"link"} error={undefined} />}
+        {getCourseStructure.isLoading && <Spinner variant={"medium"} />}
+        {getCourseStructure.isSuccess && (
+          <ManageCourseStructure
+            courseStructure={getCourseStructure.data}
+            refetch={getCourseStructure.refetch}
+          />
+        )}
       </div>
-
-      <DebugModal data={data} />
     </Layout>
   )
 }
